refactor: move logo within home
